+++ examples/nuxt/server/error.ts
@@ -1,0 +1,3 @@
+import { fedifyErrorHandler } from "@fedify/nuxt";
+
+export default fedifyErrorHandler;