+++ examples/nuxt/server/store.ts
@@ -1,0 +1,4 @@
+import type { Person } from "@fedify/vocab";
+
+export const keyPairsStore = new Map<string, CryptoKeyPair[]>();
+export const relationStore = new Map<string, Person>();